Update framework config to latest

+++ nix/mkNeovim.nix
@@ -1,18 +1,19 @@
 # Function for creating a Neovim derivation
 {
-  pkgs,
   lib,
   stdenv,
+  sqlite,
+  git,
+  neovim-unwrapped,
   # Set by the overlay to ensure we use a compatible version of `wrapNeovimUnstable`
-  pkgs-wrapNeovim ? pkgs,
+  wrapNeovimUnstable,
+  neovimUtils,
 }:
 with lib;
 {
   # NVIM_APPNAME - Defaults to 'nvim' if not set.
   # If set to something else, this will also rename the binary.
   appName ? null,
-  # The Neovim package to wrap
-  neovim-unwrapped ? pkgs-wrapNeovim.neovim-unwrapped,
   plugins ? [ ], # List of plugins
   # List of dev plugins (will be bootstrapped) - useful for plugin developers
   # { name = <plugin-name>; url = <git-url>; }
@@ -33,9 +34,10 @@ with lib;
   # You probably don't want to create vi or vim aliases
   # if the appName is something different than "nvim"
   # Add a "vi" binary to the build output as an alias?
-  viAlias ? appName == null || appname == "nvim",
+  viAlias ? appName == null || appName == "nvim",
   # Add a "vim" binary to the build output as an alias?
-  vimAlias ? appName == null || appname == "nvim",
+  vimAlias ? appName == null || appName == "nvim",
+  wrapRc ? true,
 }:
 let
   # This is the structure of a plugin definition.
@@ -47,27 +49,12 @@ let
     # set to `true`, it is installed in the 'opt' packpath, and can be lazy loaded with
     # ':packadd! {plugin-name}
     optional = false;
-    runtime = { };
   };
 
-  externalPackages = extraPackages ++ (optionals withSqlite [ pkgs.sqlite ]);
+  externalPackages = extraPackages ++ (optionals withSqlite [ sqlite ]);
 
   # Map all plugins to an attrset { plugin = <plugin>; config = <config>; optional = <tf>; ... }
   normalizedPlugins = map (x: defaultPlugin // (if x ? plugin then x else { plugin = x; })) plugins;
-
-  # This nixpkgs util function creates an attrset
-  # that pkgs.wrapNeovimUnstable uses to configure the Neovim build.
-  neovimConfig = pkgs-wrapNeovim.neovimUtils.makeNeovimConfig {
-    inherit
-      extraPython3Packages
-      withPython3
-      withRuby
-      withNodeJs
-      viAlias
-      vimAlias
-      ;
-    plugins = normalizedPlugins;
-  };
 
   # This uses the ignoreConfigRegexes list to filter
   # the nvim directory
@@ -103,11 +90,17 @@ let
     '';
 
     installPhase = ''
-      cp -r after $out/after
-      rm -r after
       cp -r lua $out/lua
       rm -r lua
-      cp -r * $out/nvim
+      # Copy nvim/after only if it exists
+      if [ -d "after" ]; then
+          cp -r after $out/after
+          rm -r after
+      fi
+      # Copy rest of nvim/ subdirectories only if they exist
+      if [ ! -z "$(ls -A)" ]; then
+          cp -r -- * $out/nvim
+      fi
     '';
   };
 
@@ -115,53 +108,56 @@ let
   # It wraps the user init.lua, prepends the lua lib directory to the RTP
   # and prepends the nvim and after directory to the RTP
   # It also adds logic for bootstrapping dev plugins (for plugin developers)
-  initLua =
+  initLua = ''
+    -- prepend lua directory
+    vim.opt.rtp:prepend('${nvimRtp}/lua')
+  ''
+  # Wrap init.lua
+  + (builtins.readFile ../nvim/init.lua)
+  # Bootstrap/load dev plugins
+  + optionalString (devPlugins != [ ]) (
     ''
-      vim.loader.enable()
-      -- prepend lua directory
-      vim.opt.rtp:prepend('${nvimRtp}/lua')
+      local dev_pack_path = vim.fn.stdpath('data') .. '/site/pack/dev'
+      local dev_plugins_dir = dev_pack_path .. '/opt'
+      local dev_plugin_path
     ''
-    # Wrap init.lua
-    + (builtins.readFile ../nvim/init.lua)
-    # Bootstrap/load dev plugins
-    + optionalString (devPlugins != [ ]) (
-      ''
-        local dev_pack_path = vim.fn.stdpath('data') .. '/site/pack/dev'
-        local dev_plugins_dir = dev_pack_path .. '/opt'
-        local dev_plugin_path
-      ''
-      + strings.concatMapStringsSep "\n" (plugin: ''
-        dev_plugin_path = dev_plugins_dir .. '/${plugin.name}'
-        if vim.fn.empty(vim.fn.glob(dev_plugin_path)) > 0 then
-          vim.notify('Bootstrapping dev plugin ${plugin.name} ...', vim.log.levels.INFO)
-          vim.cmd('!${pkgs.git}/bin/git clone ${plugin.url} ' .. dev_plugin_path)
-        end
-        vim.cmd('packadd! ${plugin.name}')
-      '') devPlugins
-    )
-    # Prepend nvim and after directories to the runtimepath
-    # NOTE: This is done after init.lua,
-    # because of a bug in Neovim that can cause filetype plugins
-    # to be sourced prematurely, see https://github.com/neovim/neovim/issues/19008
-    # We prepend to ensure that user ftplugins are sourced before builtin ftplugins.
-    + ''
-      vim.opt.rtp:prepend('${nvimRtp}/nvim')
-      vim.opt.rtp:prepend('${nvimRtp}/after')
-    '';
+    + strings.concatMapStringsSep "\n" (plugin: ''
+      dev_plugin_path = dev_plugins_dir .. '/${plugin.name}'
+      if vim.fn.empty(vim.fn.glob(dev_plugin_path)) > 0 then
+        vim.notify('Bootstrapping dev plugin ${plugin.name} ...', vim.log.levels.INFO)
+        vim.cmd('!${git}/bin/git clone ${plugin.url} ' .. dev_plugin_path)
+      end
+      vim.cmd('packadd! ${plugin.name}')
+    '') devPlugins
+  )
+  # Prepend nvim and after directories to the runtimepath
+  # NOTE: This is done after init.lua,
+  # because of a bug in Neovim that can cause filetype plugins
+  # to be sourced prematurely, see https://github.com/neovim/neovim/issues/19008
+  # We prepend to ensure that user ftplugins are sourced before builtin ftplugins.
+  + ''
+    vim.opt.rtp:prepend('${nvimRtp}/nvim')
+    vim.opt.rtp:prepend('${nvimRtp}/after')
+  '';
 
   # Add arguments to the Neovim wrapper script
-  extraMakeWrapperArgs = builtins.concatStringsSep " " (
-    # Set the NVIM_APPNAME environment variable
-    (optional (
-      appName != "nvim" && appName != null && appName != ""
-    ) ''--set NVIM_APPNAME "${appName}"'')
-    # Add external packages to the PATH
-    ++ (optional (externalPackages != [ ]) ''--prefix PATH : "${makeBinPath externalPackages}"'')
-    # Set the LIBSQLITE_CLIB_PATH if sqlite is enabled
-    ++ (optional withSqlite ''--set LIBSQLITE_CLIB_PATH "${pkgs.sqlite.out}/lib/libsqlite3.so"'')
-    # Set the LIBSQLITE environment variable if sqlite is enabled
-    ++ (optional withSqlite ''--set LIBSQLITE "${pkgs.sqlite.out}/lib/libsqlite3.so"'')
-  );
+  extraMakeWrapperArgs =
+    let
+      sqliteLibExt = stdenv.hostPlatform.extensions.sharedLibrary;
+      sqliteLibPath = "${sqlite.out}/lib/libsqlite3${sqliteLibExt}";
+    in
+    builtins.concatStringsSep " " (
+      # Set the NVIM_APPNAME environment variable
+      (optional (
+        appName != "nvim" && appName != null && appName != ""
+      ) ''--set NVIM_APPNAME "${appName}"'')
+      # Add external packages to the PATH
+      ++ (optional (externalPackages != [ ]) ''--prefix PATH : "${makeBinPath externalPackages}"'')
+      # Set the LIBSQLITE_CLIB_PATH if sqlite is enabled
+      ++ (optional withSqlite ''--set LIBSQLITE_CLIB_PATH "${sqliteLibPath}"'')
+      # Set the LIBSQLITE environment variable if sqlite is enabled
+      ++ (optional withSqlite ''--set LIBSQLITE "${sqliteLibPath}"'')
+    );
 
   luaPackages = neovim-unwrapped.lua.pkgs;
   resolvedExtraLuaPackages = extraLuaPackages luaPackages;
@@ -181,21 +177,21 @@ let
       }"'';
 
   # wrapNeovimUnstable is the nixpkgs utility function for building a Neovim derivation.
-  neovim-wrapped = pkgs-wrapNeovim.wrapNeovimUnstable neovim-unwrapped (
-    neovimConfig
-    // {
-      luaRcContent = initLua;
-      wrapperArgs =
-        escapeShellArgs neovimConfig.wrapperArgs
-        + " "
-        + extraMakeWrapperArgs
-        + " "
-        + extraMakeWrapperLuaCArgs
-        + " "
-        + extraMakeWrapperLuaArgs;
-      wrapRc = true;
-    }
-  );
+  neovim-wrapped = wrapNeovimUnstable neovim-unwrapped {
+    inherit
+      extraPython3Packages
+      withPython3
+      withRuby
+      withNodeJs
+      viAlias
+      vimAlias
+      ;
+    plugins = normalizedPlugins;
+
+    luaRcContent = initLua;
+    wrapperArgs = extraMakeWrapperArgs + " " + extraMakeWrapperLuaCArgs + " " + extraMakeWrapperLuaArgs;
+    wrapRc = wrapRc;
+  };
 
   isCustomAppName = appName != null && appName != "nvim";
 in
@@ -206,4 +202,5 @@ neovim-wrapped.overrideAttrs (oa: {
     + lib.optionalString isCustomAppName ''
       mv $out/bin/nvim $out/bin/${lib.escapeShellArg appName}
     '';
+  meta.mainProgram = if isCustomAppName then appName else oa.meta.mainProgram;
 })
